cmd, nand: add an option to disable the verification when writing in raw mode
Modern NANDs do not guarantee that data written in raw mode will not
contain bitflips just after writing it. This is fine since the number of
bitflips should be rather low and thus fixable by the ECC engine, but
since we are reading data in raw mode to verify if they match the input
data we cannot prevent failures if some bits are flipped.
The option of reading data in standard mode is not acceptable, since one
of the usage of raw mode is to allow flashing images that do not respect
the standard NAND page layout or the default ECC config (this is the case
on Allwinner platforms, where the ROM code tests several hardcoded
configs, which are not necessarily matching the NAND characteristics).

Add an extension to the nand write.raw command allowing one to disable
the verification step.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>